Move & update import to all Setters components
