+++ src/components/ArticleSet.vue
@@ -1,20 +1,20 @@
 <template>
-  <CardElt id="product-set">
+  <CardElt id="article-set">
     <template #header>
       <h2>
-        <i class="fa-regular fa-lightbulb fa-lg"></i>
-        {{ val.PRODUCT_MANAGER }}
+        <i class="fa-regular fa-pen-to-square fa-lg"></i>
+        {{ val.ARTICLE_MANAGER }}
       </h2>
     </template>
 
     <template #body>
       <form enctype="multipart/form-data">
-        <ListElt :items="val.PRODUCT_FORM">
+        <ListElt :items="val.ARTICLE_FORM">
 
           <template #item-1>
             <FieldElt id="name"
-              v-model:value="name"
-              @keyup.enter="createProduct()"
+              v-model:value="name" 
+              @keyup.enter="createArticle()"
               :info="val.INFO_NAME">
               <template #legend>{{ val.LEGEND_NAME }}</template>
               <template #label>{{ val.LABEL_NAME }}</template>
@@ -22,16 +22,16 @@
           </template>
 
           <template #item-2>
-            <label for="description">{{ val.LEGEND_DESCRIPTION }}</label>
-            <Editor id="description"
+            <label for="text">{{ val.LEGEND_TEXT }}</label>
+            <Editor id="text"
               :api-key="val.TINY_KEY"
-              v-model="description"
-              @keyup.enter="createProduct()"
+              v-model="text"
+              @keyup.enter="createArticle()"
               :init="val.TINY_INIT"/>
           </template>
 
           <template #item-3>
-            <FieldElt id="image"
+            <FieldElt id="image" 
               type="file"
               v-model:value="image"
               :info="val.INFO_IMAGE">
@@ -44,6 +44,7 @@
             <FieldElt id="alt"
               type="textarea"
               v-model:value="alt"
+              @keyup.enter="createArticle()"
               :info="val.INFO_ALT">
               <template #legend>{{ val.LEGEND_ALT }}</template>
               <template #label>{{ val.LABEL_ALT }}</template>
@@ -51,48 +52,34 @@
           </template>
 
           <template #item-5>
-            <FieldElt id="price"
-              type="number"
-              v-model:value="price"
-              @keyup.enter="createProduct()"
-              :info="val.INFO_PRICE"
-              :min="val.PRICE_MIN"
-              :max="val.PRICE_MAX">
-              <template #legend>{{ val.LEGEND_PRICE }}</template>
-              <template #label>{{ val.LABEL_PRICE }}</template>
-            </FieldElt>
-          </template>
+              <FieldElt id="url"
+                v-model:value="url"
+                @keyup.enter="updateArticle()"
+                :info="val.INFO_URL">
+                <template #legend>{{ val.LEGEND_URL }}</template>
+                <template #label>{{ val.LABEL_URL }}</template>
+              </FieldElt>
+            </template>
 
           <template #item-6>
-            <FieldElt id="options"
-              type="textarea"
-              v-model:value="options"
-              @keyup.enter="createProduct()"
-              :info="val.INFO_OPTIONS"
-              :max="100">
-              <template #legend>{{ val.LEGEND_OPTIONS }}</template>
-              <template #label>{{ val.LABEL_OPTIONS }}</template>
-            </FieldElt>
-          </template>
-
-          <template #item-7>
             <FieldElt id="cat"
               type="select"
-              :list="val.CATS_PRODUCT"
+              :list="val.CATS_ARTICLE"
               v-model:value="cat"
-              @keyup.enter="createProduct()"
+              @keyup.enter="createArticle()"
               :info="val.INFO_CAT">
               <template #legend>{{ val.LEGEND_CAT }}</template>
               <template #label>{{ val.LABEL_CAT }}</template>
             </FieldElt>
           </template>
         </ListElt>
+        <br>
 
         <BtnElt type="button"
-          @click="createProduct()" 
+          @click="createArticle()" 
           class="btn-green"
           :content="val.CONTENT_CREATE"
-          :title="val.TITLE_PRODUCT">
+          :title="val.TITLE_ARTICLE">
           <template #btn>
             <i class="fa-solid fa-square-plus fa-lg"></i>
           </template>
@@ -103,61 +90,59 @@
 </template>
 
 <script>
-import BtnElt from "../elements/BtnElt"
-import CardElt from "../elements/CardElt"
-import FieldElt from "../elements/FieldElt"
-import ListElt from "../elements/ListElt"
+import BtnElt from "./BtnElt"
+import CardElt from "./CardElt"
+import FieldElt from "./FieldElt"
+import ListElt from "./ListElt"
 
-import { checkRange } from "../../services/checkers"
-import { postData } from "../../services/fetchers"
-import { setError } from "../../services/setters"
+import { checkRange, checkRegex, postData, setError } from "../app/services"
 
 import Editor from "@tinymce/tinymce-vue"
 
 export default {
-  name: "ProductSet",
+  name: "ArticleSet",
   components: { BtnElt, CardElt, FieldElt, ListElt, Editor },
   props: ["token", "val"],
   data() {
     return {
       name: "",
-      description:"",
+      text:"",
       image: "",
       alt: "",
-      price: null,
-      options: [],
+      url: "",
       cat: ""
     }
   },
 
   methods: {
     /**
-     * ? CREATE PRODUCT
-     * * Create a product by sending a POST request to the server.
+     * ? CREATE ARTICLE
+     * * Creates an article by sending a POST request to the server with the provided data.
      */
-    createProduct() {
-      const { ALERT_CREATED, ALERT_IMG, API_URL, CAT_PRODUCT, CHECK_STRING, TEXT_MAX, TEXT_MIN } = this.val;
+    createArticle() {
+      const { ALERT_CREATED, ALERT_IMG, API_URL, CAT_ARTICLE, CHECK_STRING, CHECK_URL, REGEX_URL, TEXT_MIN, TEXT_MAX } = this.val;
 
-      if (this.price < 1) this.price = 1;
-      if (this.cat === "") this.cat = CAT_PRODUCT;
+      if (this.url.startsWith("http")) this.url = this.url.split('//')[1];
+      if (this.cat === "") this.cat = CAT_ARTICLE;
 
       const IS_NAME_CHECKED = checkRange(this.name, CHECK_STRING);
-      const IS_DESC_CHECKED = checkRange(this.description, CHECK_STRING, TEXT_MIN, TEXT_MAX);
+      const IS_TEXT_CHECKED = checkRange(this.text, CHECK_STRING, TEXT_MIN, TEXT_MAX);
       const IS_ALT_CHECKED  = checkRange(this.alt, CHECK_STRING);
 
-      if (IS_NAME_CHECKED && IS_DESC_CHECKED && IS_ALT_CHECKED) {
+      const IS_URL_CHECKED = this.url ? checkRegex(this.url, CHECK_URL, REGEX_URL) : true;
+
+      if (IS_NAME_CHECKED && IS_TEXT_CHECKED && IS_ALT_CHECKED && IS_URL_CHECKED) {
         const img = document.getElementById("image")?.files[0];
 
         if (img !== undefined) {
-          const URL   = `${API_URL}/products`;
+          const URL   = `${API_URL}/articles`;
           const data  = new FormData();
 
           data.append("name", this.name);
-          data.append("description", this.description);
+          data.append("text", this.text);
           data.append("image", img);
           data.append("alt", this.alt);
-          data.append("price", this.price);
-          data.append("options", this.options);
+          data.append("url", this.url);
           data.append("cat", this.cat);
 
           postData(URL, data, this.token)
