+++ src/components/GallerySet.vue
@@ -113,16 +113,13 @@
 </template>
 
 <script>
-import BtnElt from "../elements/BtnElt"
-import CardElt from "../elements/CardElt"
-import FieldElt from "../elements/FieldElt"
-import ListElt from "../elements/ListElt"
-import TableElt from "../elements/TableElt"
+import BtnElt from "./BtnElt"
+import CardElt from "./CardElt"
+import FieldElt from "./FieldElt"
+import ListElt from "./ListElt"
+import TableElt from "./TableElt"
 
-import { checkRange } from "../../services/checkers"
-import { deleteData, postData, putData } from "../../services/fetchers"
-import { getItemName } from "../../services/getters"
-import { setError } from "../../services/setters"
+import { checkRange, deleteData, getItemName, postData, putData, setError } from "../app/services"
 
 export default {
   name: "GallerySet",
