+++ src/components/ImageSet.vue
@@ -120,16 +120,14 @@
 </template>
 
 <script>
-import BtnElt from "../elements/BtnElt"
-import CardElt from "../elements/CardElt"
-import FieldElt from "../elements/FieldElt"
-import ListElt from "../elements/ListElt"
-import MediaElt from "../elements/MediaElt"
-import TableElt from "../elements/TableElt"
+import BtnElt from "./BtnElt"
+import CardElt from "./CardElt"
+import FieldElt from "./FieldElt"
+import ListElt from "./ListElt"
+import MediaElt from "./MediaElt"
+import TableElt from "./TableElt"
 
-import { checkRange } from "../../services/checkers"
-import { deleteData, putData, postData } from "../../services/fetchers"
-import { setError } from "../../services/setters"
+import { checkRange, deleteData, putData, postData, setError } from "../app/services"
 
 export default {
   name: "ImageSet",
