+++ src/components/ProjectSet.vue
@@ -199,17 +199,14 @@
 </template>
 
 <script>
-import BtnElt from "../elements/BtnElt"
-import CardElt from "../elements/CardElt"
-import FieldElt from "../elements/FieldElt"
-import ListElt from "../elements/ListElt"
-import MediaElt from "../elements/MediaElt"
-import TableElt from "../elements/TableElt"
+import BtnElt from "./BtnElt"
+import CardElt from "./CardElt"
+import FieldElt from "./FieldElt"
+import ListElt from "./ListElt"
+import MediaElt from "./MediaElt"
+import TableElt from "./TableElt"
 
-import { checkRange, checkRegex } from "../../services/checkers"
-import { deleteData, postData, putData } from "../../services/fetchers"
-import { getItemName } from "../../services/getters"
-import { setError } from "../../services/setters"
+import { checkRange, checkRegex, deleteData, getItemName, postData, putData, setError } from "../app/services"
 
 export default {
   name: "ProjectSet",
