+++ src/components/UserSet.vue
@@ -95,16 +95,13 @@
 </template>
 
 <script>
-import BtnElt from "../elements/BtnElt"
-import CardElt from "../elements/CardElt"
-import FieldElt from "../elements/FieldElt"
-import MediaElt from "../elements/MediaElt"
-import TableElt from "../elements/TableElt"
+import BtnElt from "./BtnElt"
+import CardElt from "./CardElt"
+import FieldElt from "./FieldElt"
+import MediaElt from "./MediaElt"
+import TableElt from "./TableElt"
 
-import { checkRange, checkRegex } from "../../services/checkers"
-import { deleteData, putData } from "../../services/fetchers"
-import { getItemName } from "../../services/getters"
-import { setError } from "../../services/setters"
+import { checkRange, checkRegex, deleteData, getItemName, putData, setError } from "../app/services"
 
 export default {
   name: "UserSet",
